clang server errors

+++ src/server.c
@@ -27,7 +27,7 @@ void handle_receive(int client_socket_list[MAX_CLIENTS],
                     char buffer[MAX_MESSAGE_LENGTH], int current_client) {
   // number of bytes received; 0=client disconnect, <0=error, >0=message
   int bytes_received = (int) recv(client_socket_list[current_client], buffer,
-                            MAX_MESSAGE_LENGTH * 4, 0);
+                            (size_t) (MAX_MESSAGE_LENGTH * 4), 0);
   // if client is sending a message
   if (bytes_received > 0) {
     buffer[bytes_received] = '\0';
